==> scripts/dDNA_process/0_runQC.Rmd <==
---
title: "Run QC - Demultiplexed FastQs"
date: '2022-09-27'
author: M Fisher
output: 
  html_document:
    toc: yes
---

# Description

Graph the distribution of read depths from MiSeq sequencing output, and look at base pair quality along the length of the sequence.

Dependencies: Run the bash script `00_get_fastq_read_depth.sh` from the command line (or by double clicking it) to determine the number of reads in the fastq files (counts the lines and divides by 4). 

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = FALSE)

library(here)
library(tidyverse)
library(magrittr)
library(ggplot2)
library(readr)
library(readxl)
library(janitor)

## to install dada2
# if (!requireNamespace("BiocManager", quietly = TRUE))
#     install.packages("BiocManager")
# BiocManager::install("dada2", version = "3.14")
library(dada2)
```
<br>

# Run 1 

This sequencing run was completed at UCLA.

The sequence quality section requires the location of the fastq files:
```{r echo=TRUE}
fastq_dir <- "data/raw/run_1"
```
<br>


The read depth information for run 2 was provided in a "demultiplexing report" excel document by the UCLA TCGB. 
```{r}
dat <- readxl::read_excel(here::here("data","miseq_run_stats","Run1_DemultiplexingStatsUCLA.xlsx"))
head(dat)
```
<br>


<br>

## Read depth (MiSeq Stats)
<br>
```{r}
dat %<>% filter(!is.na(name2)) %>% filter(name3 != "all") %>%
  dplyr::select(-c(`flowcell-id`,PerfectBarcodeCount,OneMismatchBarcodeCount))

dat %<>% rename(miseq_id=name2,barcode=name3,nreads=BarcodeCount) %>% dplyr::select(-name,-number) %>%
  separate(miseq_id, into=c("sample_name","site","sample_id"), sep="_", remove=FALSE)


dat %<>% filter(!is.na(sample_id)) %>% separate(sample_id, into=c("sample_id","rep"), remove=TRUE, sep="-")
```

```{r eval=FALSE}
write_csv(dat, file=here("data","miseq_run_stats","Run1_DemultiplexingStats.csv"))
```
<br>

```{r fig.width=6, fig.height=3, warning=FALSE, message=FALSE, echo=FALSE}
fig1a <- ggplot(dat, aes(x=nreads/1000)) + geom_histogram() +
  geom_vline(aes(xintercept=140), lty=2) +
  labs(x="Reads per Sample (x1000)", y="No. Samples",
       title="All Samples Run 1",
       subtitle="MiSeq 2x300") +
  theme_classic() + theme(axis.text.x=element_text(size=12),
                          axis.title=element_text(size=14),
                          title=element_text(size=16))
fig1a
```

*I believe these are total forward and reverse reads, so a cutoff of 70,000 reads per direction would be x=140 (vertical dashed line above)*

<br>

```{r fig.height=4, fig.width=6}
fig2 <- ggplot(dat, aes(x=site, y=nreads/1000, fill=site)) +
  geom_boxplot(alpha=0.8) +
  geom_hline(aes(yintercept=70), lty=2, col="grey45") +
  geom_hline(aes(yintercept=140), lty=2, col="grey45") +
  geom_point(aes(fill=site), alpha=0.4) + 
  labs(y="Reads per Sample (x1000)", x="Site",
       title="All Samples Run 1",
       subtitle="MiSeq 2x300") +
  theme_bw() + theme(axis.text.x=element_text(size=11, angle=90, hjust=1,vjust=0.5),
                     axis.title=element_text(size=12),
                     strip.text=element_text(size=11),
                     title=element_text(size=14),
                     legend.title=element_blank())
fig2
```
<br>
<br>

Which samples have zero reads? *None!*
```{r}
dat %>% filter(nreads==0)
```
<br>

Which samples have fewer than 20,000 reads (10,000 in each direction)
```{r}
dat %>% filter(nreads <= 20000)
```
<br>

Which samples had a very large number of reads? *CLAY_05*
```{r}
dat %>% filter(nreads > 400000)
```
<br>
<br>

## Read depth (Fastq)


This is the read depth determined using the fastq files.

```{r message=FALSE, warning=FALSE}
dat1 <- read_csv(here(fastq_dir,"run1_readcounts.csv"), col_names=FALSE ) %>%
  rename("file"=X1, "nreads"=X2)

dat1 %<>% mutate(direction=ifelse(grepl("R1", file), "forward","reverse"),
  file=ifelse(grepl("R1", file), str_remove(file,"_L001_R1_001.fastq.gz"),str_remove(file,"_L001_R2_001.fastq.gz")))
dat1 %<>% separate(file, into=c("temp1","sample_num"), sep="_")
dat1 %<>% separate(temp1, into=c("temp2", "site","crab","tech"), sep="-") %>%
  dplyr::select(-temp2) %>%
  mutate(crab=ifelse(grepl("Control",site) | grepl("NONE",site), "",crab)) %>%
  unite(col=sample_label, crab,tech, sep="-")
```

```{r fig.height=4, fig.width=6}
fig2.1 <- ggplot(dat1, aes(x=site, y=nreads/1000, fill=site)) +
  geom_boxplot(alpha=0.8) +
  # geom_hline(aes(yintercept=70), lty=2, col="grey45") +
  # geom_hline(aes(yintercept=140), lty=2, col="grey45") +
  geom_point(aes(fill=site), alpha=0.4) + 
  facet_grid(cols=vars(direction)) +
  labs(y="Reads per Sample (x1000)", x="Group",
       title="All Samples Run 1",
       subtitle="MiSeq 2x300") +
  theme_bw() + theme(axis.text.x=element_text(size=11, angle=90, hjust=1,vjust=0.5),
                     axis.title=element_text(size=12),
                     strip.text=element_text(size=11),
                     title=element_text(size=14),
                     legend.title=element_blank())
fig2.1
```


## Check read quality

```{r include=FALSE}
## forward file names
ffilenames <- list.files(path = here(fastq_dir), 
                         pattern="*R1_001.fastq.gz")
## reverse file names
rfilenames <- list.files(path = here(fastq_dir), 
                         pattern="*R2_001.fastq.gz")
file.metadata <- bind_cols("file1"=ffilenames,"file2"=rfilenames) %>%
  mutate(file1=paste0(here(fastq_dir),"/",file1),
         file2=paste0(here(fastq_dir),"/",file2))

# grab forward and reverse file names for four unique samples/replicates. This assumes each sample has 3 replicates (a/b/c).
rows_to_subset <- seq(from=1,to=dim(file.metadata)[1]-3,by=3); rows_to_subset=sample(rows_to_subset, size=4)
subset <- file.metadata %>% slice(rows_to_subset)
```
<br>
```{r eval=FALSE}
file.metadata %>% pull(file1) %>%
  plotQualityProfile(., aggregate=TRUE)
file.metadata %>% pull(file2) %>%
 plotQualityProfile(., aggregate=TRUE)
```

Look into a subset of the samples. 
```{r}
subset %>% pull(file1) %>%
  plotQualityProfile(.)
subset %>% pull(file2) %>%
  plotQualityProfile(.)
```
<br>


# Run 2

This sequencing run was completed at the Center for Environmental Genomics, on the Kelly Lab MiSeq

The sequence quality section requires the location of the fastq files:
```{r echo=TRUE}
fastq_dir2 <- "data/raw/run_2"
```
<br>


## Read Depth (FastQ)

This is the read depth determined using the fastq files.

```{r message=FALSE, warning=FALSE}
dat2 <- read_csv(here(fastq_dir2,"run2_readcounts.csv"), col_names=FALSE ) %>%
  rename("file"=X1, "nreads"=X2)

dat2 %<>% mutate(direction=ifelse(grepl("R1", file), "forward","reverse"),
  file=ifelse(grepl("R1", file), str_remove(file,"_L001_R1_001.fastq.gz"),str_remove(file,"_L001_R2_001.fastq.gz")))
dat2 %<>% separate(file, into=c("site","sample_label","sample_num"), sep="_") %>%
  mutate(sample_num=ifelse(grepl("Control",site) | grepl("NONE",site), sample_label,sample_num),
         sample_label=ifelse(grepl("Control",site) | grepl("NONE",site), NA, sample_label))
```

```{r fig.height=4, fig.width=6}
fig2.2 <- ggplot(dat2, aes(x=site, y=nreads/1000, fill=site)) +
  geom_boxplot(alpha=0.8) +
  # geom_hline(aes(yintercept=70), lty=2, col="grey45") +
  # geom_hline(aes(yintercept=140), lty=2, col="grey45") +
  geom_point(aes(fill=site), alpha=0.4) + 
  facet_grid(cols=vars(direction)) +
  labs(y="Reads per Sample (x1000)", x="Group",
       title="All Samples Run 2",
       subtitle="MiSeq 2x300") +
  theme_bw() + theme(axis.text.x=element_text(size=11, angle=90, hjust=1,vjust=0.5),
                     axis.title=element_text(size=12),
                     strip.text=element_text(size=11),
                     title=element_text(size=14),
                     legend.title=element_blank())
fig2.2
```

Good that there aren't reads in the "NONE" samples, which were indices that I didn't include in this run; and in the Negative Control.
The Positive Control has a lot of reads, hopefully that's all roo. 



```{r fig.height=4, fig.width=6}
fig3.2 <- ggplot(dat2 %>% filter(!(grepl("Control",site)) & !(grepl("NONE",site))), aes(x=site, y=nreads/1000, fill=site)) +
  geom_boxplot(alpha=0.8) +
  # geom_hline(aes(yintercept=70), lty=2, col="grey45") +
  # geom_hline(aes(yintercept=140), lty=2, col="grey45") +
  geom_point(aes(fill=site), alpha=0.4) + 
  facet_grid(cols=vars(direction)) +
  labs(y="Reads per Sample (x1000)", x="Site",
       title="All Samples Run 2",
       subtitle="MiSeq 2x300") +
  theme_bw() + theme(axis.text.x=element_text(size=11, angle=90, hjust=1,vjust=0.5),
                     axis.title=element_text(size=12),
                     strip.text=element_text(size=11),
                     title=element_text(size=14),
                     legend.title=element_blank())
fig3.2
```

## Check read quality

```{r include=FALSE}
## forward file names
ffilenames <- list.files(path = here(fastq_dir2), 
                         pattern="*R1_001.fastq.gz")
## reverse file names
rfilenames <- list.files(path = here(fastq_dir2), 
                         pattern="*R2_001.fastq.gz")
file2.metadata <- bind_cols("file1"=ffilenames,"file2"=rfilenames) %>%
  mutate(file1=paste0(here(fastq_dir2),"/",file1),
         file2=paste0(here(fastq_dir2),"/",file2))

# grab forward and reverse file names for four unique samples/replicates. This assumes each sample has 3 replicates (a/b/c).
rows_to_subset2 <- seq(from=1,to=dim(file2.metadata)[1]-3,by=3); rows_to_subset2=sample(rows_to_subset2, size=4)
subset2 <- file2.metadata %>% slice(rows_to_subset2)
```
<br>

Look into a subset of the samples. 
```{r}
subset2 %>% pull(file1) %>%
  plotQualityProfile(.)
subset2 %>% pull(file2) %>%
  plotQualityProfile(.)
```
<br>




# Read Counts across Runs


```{r}
read.dat <- bind_rows(dat1 %>% mutate(MiSeqRun="1"),dat2 %>% mutate(MiSeqRun="2")) %>%
  filter(!(grepl("NONE",site))) %>%
  mutate(site=ifelse(site=="NegativeControl","Negative",
                     ifelse(site=="PositiveControl","Positive",site)))

read.dat$site <- factor(read.dat$site, levels=c("CLAY","KAY","MARPT","SAMT","SIN","PBNERR","Positive","Negative"))

ggplot(read.dat %>% filter(!(grepl("Control",site)) & !(grepl("NONE",site))), aes(x=site, y=nreads/1000, fill=MiSeqRun)) +
  geom_boxplot(alpha=0.8) +
  # geom_hline(aes(yintercept=70), lty=2, col="grey45") +
  # geom_hline(aes(yintercept=140), lty=2, col="grey45") +
  geom_point(alpha=0.4) + 
  facet_grid(cols=vars(direction)) +
  labs(y="Reads per Sample (x1000)", x="Site",
       title="All Samples All Runs",
       subtitle="MiSeq 2x300") +
  theme_bw() + theme(axis.text.x=element_text(size=11, angle=90, hjust=1,vjust=0.5),
                     axis.title=element_text(size=12),
                     strip.text=element_text(size=11),
                     title=element_text(size=14))
```
